requirements.txt
# Django Core
Django>=5.1,<5.2
Pillow>=10.0,<12.2
djangorestframework>=3.15,<3.16
django-filter>=24.0

# Authentication JWT
djangorestframework-simplejwt>=5.3,<6.0

# PostgreSQL + PostGIS
psycopg2-binary>=2.9,<3.0

# Celery + Redis
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Environment variables
python-decouple>=3.8,<4.0

# Geospatial utilities (for Haversine when PostGIS not available)
geopy>=2.4,<3.0

# Phone number validation
django-phonenumber-field>=8.0,<9.0
phonenumbers>=8.13,<9.0

# API documentation
drf-spectacular>=0.27,<1.0

# Deployment
gunicorn>=21.2,<23.0
whitenoise[brotli]>=6.6,<7.0
dj-database-url>=2.1,<3.0

# Dev tools
django-extensions>=3.2,<4.0
